Inventory Menu, "Bites" system, mouse aiming, pickups
I created a 4-panel UI menu (which is always up) where you can click on an item and click again to place it in another slot. I made it so items have several uses before running out (I call this "bites"). Projectiles shoot in the direction of your mouse. Also, you can pick up that sphere on the ground and it will give you more "fireballs"